requirements.txt
phidata
groq
python-dotenv